Potential fix for code scanning alert no. 8: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using unsigned long arithmetic to prevent overflow. This can be achieved by casting one of the operands to unsigned long before performing the multiplication. This way, the multiplication will be done in the larger type, and the result will be correctly stored in the unsigned long variable.

Specifically, we will cast num_instances to unsigned long before multiplying it by num_axes and sizeof(FT_Fixed) on line 2080.

+++ libfreetype2-2.10.1/src/truetype/ttgxvar.c
@@ -2077,7 +2077,7 @@
                                    sizeof ( FT_Var_Axis ) );
     namedstyle_size  = ALIGN_SIZE( num_instances *
                                    sizeof ( FT_Var_Named_Style ) );
-    next_coords_size = ALIGN_SIZE( num_instances *
+    next_coords_size = ALIGN_SIZE( (unsigned long)num_instances *
                                    num_axes *
                                    sizeof ( FT_Fixed ) );
     next_name_size   = num_axes * 5;
